requirements.txt
streamlit
mysql-connector-python
pandas
pillow
prophet
